+++ azure-kusto-data/pyproject.toml
@@ -17,7 +17,6 @@ classifiers=[
 keywords=["kusto", "wrapper", "client", "azure", "data", "query"]
 packages = [
     { include = "azure/kusto/data"},
-    { include = "tests"}
 ]
 
 [tool.poetry.dependencies]
