+++ azure-kusto-data/tests_data/input/adminthenquery.json
