+++ azure-kusto-data/tests_data/input/dataframe.json
