+++ azure-kusto-data/tests_data/input/deft.json
