+++ azure-kusto-data/tests_data/input/deft_with_progressive_result.json
