+++ azure-kusto-data/tests_data/input/dynamic.json
