+++ azure-kusto-data/tests_data/input/pandas_bool.json
