+++ azure-kusto-data/tests_data/input/progressive_result.json
