+++ azure-kusto-data/tests_data/input/query_partial_results_defer_is_false.json
