+++ azure-kusto-data/tests_data/input/versionshowcommandresult.json
