+++ azure-kusto-data/tests_data/input/zero_results.json
