+++ azure-kusto-data/tests_data/test_client_request_properties.py
