+++ azure-kusto-data/tests_data/test_exceptions.py
