+++ azure-kusto-data/tests_data/test_functional.py
