+++ azure-kusto-data/tests_data/test_http_adapter_with_socket_options.py
