+++ azure-kusto-data/tests_data/test_kusto_client.py
@@ -10,7 +10,7 @@ from azure.kusto.data._cloud_settings import CloudSettings
 from azure.kusto.data.exceptions import KustoClosedError, KustoMultiApiError
 from azure.kusto.data.helpers import dataframe_from_result_table
 from azure.kusto.data.response import KustoStreamingResponseDataSet
-from tests.kusto_client_common import KustoClientTestsMixin, mocked_requests_post, get_response_first_primary_result, get_table_first_row, proxy_kcsb
+from tests_data.kusto_client_common import KustoClientTestsMixin, mocked_requests_post, get_response_first_primary_result, get_table_first_row, proxy_kcsb
 
 PANDAS = False
 try:
