+++ azure-kusto-data/tests_data/test_telemetry.py
