+++ azure-kusto-data/tests_data/test_well_known_kusto_endpoints.py
