+++ azure-kusto-ingest/tests_ingest/input/dataset.jsonz.gz
