+++ azure-kusto-ingest/tests_ingest/test_connection_string.py
