+++ azure-kusto-ingest/tests_ingest/test_ingestion_properties.py
