+++ azure-kusto-ingest/tests_ingest/test_kusto_streaming_ingest_client.py
@@ -10,6 +10,7 @@ import responses
 
 from azure.kusto.data.data_format import DataFormat
 from azure.kusto.ingest import KustoStreamingIngestClient, IngestionProperties, IngestionStatus, ManagedStreamingIngestClient
+from .common import get_input_folder_path
 
 pandas_installed = False
 try:
@@ -79,63 +80,31 @@ class TestKustoStreamingIngestClient:
         ingest_client = ingest_client_class("https://somecluster.kusto.windows.net")
         ingestion_properties = IngestionProperties(database="database", table="table", data_format=DataFormat.CSV)
 
-        # ensure test can work when executed from within directories
-        current_dir = os.getcwd()
-        path_parts = ["azure-kusto-ingest", "tests", "input", "dataset.csv"]
-        missing_path_parts = []
-        for path_part in path_parts:
-            if path_part not in current_dir:
-                missing_path_parts.append(path_part)
-
-        file_path = os.path.join(current_dir, *missing_path_parts)
+        file_path = str(get_input_folder_path("dataset.csv"))
 
         result = ingest_client.ingest_from_file(file_path, ingestion_properties=ingestion_properties)
         assert result.status == IngestionStatus.SUCCESS
 
-        path_parts = ["azure-kusto-ingest", "tests", "input", "dataset.csv.gz"]
-        missing_path_parts = []
-        for path_part in path_parts:
-            if path_part not in current_dir:
-                missing_path_parts.append(path_part)
-
-        file_path = os.path.join(current_dir, *missing_path_parts)
+        file_path = str(get_input_folder_path("dataset.csv.gz"))
 
         result = ingest_client.ingest_from_file(file_path, ingestion_properties=ingestion_properties)
         assert result.status == IngestionStatus.SUCCESS
 
         ingestion_properties = IngestionProperties(database="database", table="table", data_format=DataFormat.JSON, ingestion_mapping_reference="JsonMapping")
 
-        path_parts = ["azure-kusto-ingest", "tests", "input", "dataset.json"]
-        missing_path_parts = []
-        for path_part in path_parts:
-            if path_part not in current_dir:
-                missing_path_parts.append(path_part)
-
-        file_path = os.path.join(current_dir, *missing_path_parts)
+        file_path = str(get_input_folder_path("dataset.json"))
 
         result = ingest_client.ingest_from_file(file_path, ingestion_properties=ingestion_properties)
         assert result.status == IngestionStatus.SUCCESS
 
-        path_parts = ["azure-kusto-ingest", "tests", "input", "dataset.jsonz.gz"]
-        missing_path_parts = []
-        for path_part in path_parts:
-            if path_part not in current_dir:
-                missing_path_parts.append(path_part)
-
-        file_path = os.path.join(current_dir, *missing_path_parts)
+        file_path = str(get_input_folder_path("dataset.jsonz.gz"))
 
         result = ingest_client.ingest_from_file(file_path, ingestion_properties=ingestion_properties)
         assert result.status == IngestionStatus.SUCCESS
 
         ingestion_properties = IngestionProperties(database="database", table="table", data_format=DataFormat.TSV)
 
-        path_parts = ["azure-kusto-ingest", "tests", "input", "dataset.tsv"]
-        missing_path_parts = []
-        for path_part in path_parts:
-            if path_part not in current_dir:
-                missing_path_parts.append(path_part)
-
-        file_path = os.path.join(current_dir, *missing_path_parts)
+        file_path = str(get_input_folder_path("dataset.tsv"))
 
         result = ingest_client.ingest_from_file(file_path, ingestion_properties=ingestion_properties)
         assert result.status == IngestionStatus.SUCCESS
